S05P13C103-63 S05P13C103-62 :feat [BE] 라이브방송 시작 종료 islive변경, 물건 판매완료시 issold1로 변경

+++ backend/src/main/java/com/ssafy/api/response/dto/Product/ProductListResponseDto.java
@@ -12,7 +12,7 @@ import java.util.Set;
 public class ProductListResponseDto {
     private Long productId;
     private String userId;
-    private Integer liveId;
+    private Long liveId;
     private Long categoryId;
     private String title;
     private String description;
@@ -20,7 +20,7 @@ public class ProductListResponseDto {
     private LocalDateTime createdAt;
     private Integer viewCount;
     private Integer isSold;
-    private Integer isLive;
+    private Long isLive;
     private Image thumbnail;
